:lipstick: Optimize mac permission request UI
