+++ composeApp/src/commonMain/kotlin/com/clipevery/ClipeveryApp.kt
@@ -27,7 +27,6 @@ import com.clipevery.ui.ClipeveryTheme
 import com.clipevery.ui.HomeView
 import com.clipevery.ui.PageViewType
 import com.clipevery.ui.base.DialogService
-import com.clipevery.ui.base.MessageView
 import com.clipevery.ui.base.ToastManager
 import com.clipevery.ui.base.ToastView
 import com.clipevery.ui.devices.DeviceDetailView
@@ -105,8 +104,6 @@ fun ClipeveryWindow(hideWindow: suspend () -> Unit) {
                 ) {
                     ClipeveryContent()
                 }
-
-                MessageView()
 
                 toast?.let {
                     ToastView(toast = it) {
