+++ composeApp/src/commonMain/kotlin/com/clipevery/listener/GlobalListener.kt
@@ -1,16 +1,10 @@
 package com.clipevery.listener
 
-import com.clipevery.ui.base.ComposeMessageViewFactory
-
 interface GlobalListener {
-
-    var errorCode: Int?
 
     fun isRegistered(): Boolean
 
     fun start()
 
     fun stop()
-
-    fun getComposeMessageViewFactory(): ComposeMessageViewFactory
 }
