--- composeApp/src/commonMain/kotlin/com/clipevery/ui/base/ComposeMessageViewFactory.kt
@@ -1,11 +1,0 @@
-package com.clipevery.ui.base
-
-import androidx.compose.runtime.Composable
-
-interface ComposeMessageViewFactory {
-
-    var showMessage: Boolean
-
-    @Composable
-    fun MessageView(key: Any)
-}
