--- composeApp/src/commonMain/kotlin/com/clipevery/ui/base/MessageManager.kt
@@ -1,10 +1,0 @@
-package com.clipevery.ui.base
-
-import androidx.compose.runtime.Composable
-
-interface MessageManager {
-
-    var messageId: Int
-
-    fun getCurrentMessageView(): (@Composable () -> Unit)?
-}
