+++ composeApp/src/desktopMain/kotlin/com/clipevery/Clipevery.kt
@@ -137,13 +137,11 @@ import com.clipevery.ui.ThemeDetector
 import com.clipevery.ui.WindowsTray
 import com.clipevery.ui.base.DesktopDialogService
 import com.clipevery.ui.base.DesktopIconStyle
-import com.clipevery.ui.base.DesktopMessageManager
 import com.clipevery.ui.base.DesktopNotificationManager
 import com.clipevery.ui.base.DesktopToastManager
 import com.clipevery.ui.base.DesktopUISupport
 import com.clipevery.ui.base.DialogService
 import com.clipevery.ui.base.IconStyle
-import com.clipevery.ui.base.MessageManager
 import com.clipevery.ui.base.NotificationManager
 import com.clipevery.ui.base.ToastManager
 import com.clipevery.ui.base.UISupport
@@ -303,12 +301,11 @@ class Clipevery {
                     single<DesktopMouseListener> { DesktopMouseListener }
                     single<NativeMouseListener> { get<DesktopMouseListener>() }
                     single<ActiveGraphicsDevice> { get<DesktopMouseListener>() }
-                    single<GlobalListener> { DesktopGlobalListener(get(), get()) }
+                    single<GlobalListener> { DesktopGlobalListener(get(), get(), get(), get(), get()) }
                     single<ThemeDetector> { DesktopThemeDetector(get()) }
                     single<ClipResourceLoader> { DesktopAbsoluteClipResourceLoader }
                     single<ToastManager> { DesktopToastManager() }
                     single<NotificationManager> { DesktopNotificationManager }
-                    single<MessageManager> { DesktopMessageManager(get()) }
                     single<IconStyle> { DesktopIconStyle }
                     single<UISupport> { DesktopUISupport(get(), get()) }
                     single<ShortcutKeys> { DesktopShortcutKeys(get(), get()) }
