+++ composeApp/src/desktopMain/resources/i18n/en.properties
@@ -43,12 +43,14 @@ Do_you_trust_this_device?=Do you trust this device?
 Edit_Shortcut_Key=Edit Shortcut Key
 Empty=Empty
 Encrypted_sync=Encrypted sync
+Error_Code=Error Code
 Expiration_Cleanup=Expiration Cleanup
 FQA=FQA↗
 Failed_to_browse_File_pasteboard=Failed to browse File pasteboard
 Failed_to_open_Html_pasteboard=Failed to open Html pasteboard
 Failed_to_open_Image_pasteboard=Failed to open Image pasteboard
 Failed_to_open_Text_pasteboard=Failed to open Text pasteboard
+Failed_to_register_keyboard_listener=Failed to register keyboard listener
 Favorite=Favorite
 Favorite_Storage=Favorite Storage
 Feedback=Feedback
@@ -96,7 +98,8 @@ Port=Port
 Quit=Quit
 Remote=Remote
 Remove_Device=Remove Device
-Restart_Application=If authorized, please click to restart the application
+Restart_Application_Tip=If authorized, please click to restart the application
+Restart_Application=Restart Application
 Return=Return
 Scan=Scan
 Scroll_to_top=Scroll to top
