+++ composeApp/src/desktopMain/resources/i18n/es.properties
@@ -44,12 +44,14 @@ Do_you_trust_this_device?=Confía en este dispositivo?
 Edit_Shortcut_Key=Editar tecla de acceso rápido
 Empty=Vacío
 Encrypted_sync=Sincronización cifrada
+Error_Code=Código de error
 Expiration_Cleanup=Limpieza de expiración
 FQA=Preguntas frecuentes↗
 Failed_to_browse_File_pasteboard=Error al navegar por el portapapeles de archivo
 Failed_to_open_Html_pasteboard=Error al abrir el portapapeles Html
 Failed_to_open_Image_pasteboard=Error al abrir el portapapeles de imagen
 Failed_to_open_Text_pasteboard=Error al abrir el portapapeles de texto
+Failed_to_register_keyboard_listener=Error al registrar el escuchador de teclado
 Favorite=Favorito
 Favorite_Storage=Almacenamiento favorito
 Feedback=Comentarios
@@ -97,7 +99,8 @@ Port=Puerto
 Quit=Salir
 Remote=Remoto
 Remove_Device=Eliminar dispositivo
-Restart_Application=Si está autorizado, haga clic para reiniciar la aplicación.
+Restart_Application_Tip=Si está autorizado, haga clic para reiniciar la aplicación.
+Restart_Application=Reiniciar aplicación
 Return=Volver
 Scan=Escanear
 Scroll_to_top=Desplazarse hacia arriba
