+++ composeApp/src/desktopMain/resources/i18n/jp.properties
@@ -43,12 +43,14 @@ Do_you_trust_this_device?=このデバイスを信頼しますか？
 Edit_Shortcut_Key=ショートカットキーを編集
 Empty=空
 Encrypted_sync=暗号化同期
+Error_Code=エラーコード
 Expiration_Cleanup=有効期限クリーンアップ
 FQA=よくある問題↗
 Failed_to_browse_File_pasteboard=ファイルペーストボードをブラウズできませんでした
 Failed_to_open_Html_pasteboard=Html ペーストボードを開けませんでした
 Failed_to_open_Image_pasteboard=画像ペーストボードを開けませんでした
 Failed_to_open_Text_pasteboard=テキストペーストボードを開けませんでした
+Failed_to_register_keyboard_listener=キーボードリスナーの登録に失敗しました
 Favorite=お気に入り
 Favorite_Storage=お気に入りのストレージ
 Feedback=フィードバック
@@ -96,7 +98,8 @@ Port=ポート
 Quit=終了
 Remote=リモート
 Remove_Device=デバイスを削除
-Restart_Application=許可されている場合は、クリックしてアプリケーションを再起動してください
+Restart_Application_Tip=許可されている場合は、クリックしてアプリケーションを再起動してください
+Restart_Application=アプリケーションを再起動
 Return=戻る
 Scan=スキャン
 Scroll_to_top=トップにスクロール
