+++ composeApp/src/desktopMain/resources/i18n/zh.properties
@@ -43,12 +43,14 @@ Do_you_trust_this_device?=您信任此设备吗？
 Edit_Shortcut_Key=编辑快捷键
 Empty=空
 Encrypted_sync=加密同步
+Error_Code=错误码
 Expiration_Cleanup=过期清理
 FQA=常见问题↗
 Failed_to_browse_File_pasteboard=无法浏览文件剪贴板
 Failed_to_open_Html_pasteboard=无法打开 Html 剪贴板
 Failed_to_open_Image_pasteboard=无法打开图片剪贴板
 Failed_to_open_Text_pasteboard=无法打开文本剪贴板
+Failed_to_register_keyboard_listener=无法注册键盘监听器
 Favorite=收藏
 Favorite_Storage=收藏存储
 Feedback=问题反馈
@@ -96,7 +98,8 @@ Port=端口
 Quit=退出
 Remote=远程
 Remove_Device=移除设备
-Restart_Application=如果已授权，请点击重启应用
+Restart_Application_Tip=如果已授权，请点击重启应用
+Restart_Application=重启应用
 Return=返回
 Scan=扫描
 Scroll_to_top=滚动到顶部
