+++ composeApp/src/desktopMain/swift/MacosApi.swift
@@ -1,4 +1,5 @@
 import AppKit
+import ApplicationServices
 import Cocoa
 import Security
 
@@ -282,3 +283,13 @@ public func simulatePasteCommand(keyCodesPointer: UnsafePointer<Int32>, count: I
         }
     }
 }
+
+@_cdecl("checkAccessibilityPermissions")
+public func checkAccessibilityPermissions() -> Bool {
+    let checkOptionPrompt = kAXTrustedCheckOptionPrompt.takeUnretainedValue() as String
+    let options = [checkOptionPrompt: true] as CFDictionary
+    let accessEnabled = AXIsProcessTrustedWithOptions(options)
+
+    return accessEnabled
+}
+
